pair production threshold for f2_nc
