+++ src/yadism/coefficient_functions/heavy/partonic_channel.py
@@ -40,27 +40,6 @@ class NeutralCurrentBase(pc.PartonicChannel):
             return lambda: pc.RSL()
         return f
 
-    def is_below_threshold(self, z):
-        """
-        Checks if the available energy is below production threshold or not
-
-        Parameters
-        ----------
-            z : float
-                partonic momentum fraction
-
-        Returns
-        -------
-            is_below_threshold : bool
-                is the partonic energy sufficient to create the heavy quark
-                pair?
-
-        .. todo::
-            use threshold on shat or using FH's zmax?
-        """
-        shat = self.ESF.Q2 * (1 - z) / z
-        return shat <= 4 * self.m2hq
-
 
 class ChargedCurrentBase(pc.PartonicChannel):
     r"""
