+++ src/yadism/coefficient_functions/partonic_channel.py
@@ -150,6 +150,27 @@ class PartonicChannel(dict):
         """
         return f
 
+    def is_below_threshold(self, z):
+        """
+        Checks if the available energy is below production threshold or not
+
+        Parameters
+        ----------
+            z : float
+                partonic momentum fraction
+
+        Returns
+        -------
+            is_below_threshold : bool
+                is the partonic energy sufficient to create the heavy quark
+                pair?
+
+        .. todo::
+            use threshold on shat or using FH's zmax?
+        """
+        shat = self.ESF.Q2 * (1 - z) / z
+        return shat <= 4 * self.m2hq
+
     @staticmethod
     def LO():
         return None
